A fix for the problem I mentioned earlier about having multiple NVIZ viewers open.
The problem was that if multiple NVIZ viewers were open they each have a
unique name. Unfortunately the scripting send command needs to know the
name in order for the script to work. Originally this was hard set to
NVIZ. I now pick up the name when the interface starts and save it as an
environment variable which can be picked up at any time by the interface.
This seems to have solved the problem.