Date: Fri, 18 Oct 2002 12:27:02 -0300 From: Bob Covill <bcovill@tekmap.ns.ca> Subject: NVIZ update
I have been cleaning up the NVIZ off-screen code and have attached the
changes.

The cleanup includes better handling of the off-screen context and a
compile option to use a pixmap instead of a pbuffer. The pixmap would
only be usable with a lower end graphics setup that does not have direct
rendering. I have also fixed the destroy pbuffer routine so that the
togl display is returned when the animation is complete. These changes
are included in gsd_prim.c.

I have also changed the panel_kanimator.tcl so that off-screen is an
option when rendering. The destroy pbuffer routine is tied into the
animation stop routine which is executed when the animation completes or
the stop button is pressed. This should destroy the pbuffer and return
the togl context (original window).

I have also added an unrelated option to the panel_kanimator.tcl that
allows the user to set a starting frame number (default = 0). This is
useful when rendering animations in different sections, and you want to
continue numbering from where the last section left off. I have also
changed the number format from 5 digits to 4. This simply makes it more
compatible with the gimp animation scripting tools.

It is important to note that when running the off-screen rendering, the
original togl window should be minimized. If it is isn't, it can
interfere with the off-screen context. As a general rule, minimize the
window before running in off-screen mode. Also, I have found that
wireframe off-screen rendering is temperamental and requires a bit more
work.

If you have any questions or problems let me know.

--
Bob Covill

Tekmap Consulting
P.O. Box 2016
Fall River, N.S.
B2T 1K6
Canada