This is an incremental version that makes the changes to the next major version more clear. It is tested and works, but isn't really ever going to be used.
Options and Output are now displayed in tabs instead of panes

There is a new command label which shows the current command as it is edited.
It has a button for copying the command to the clipboard. (Might be worth
noting in wish 2074)

Programs reporting guisections will have a tab made for options in each
toplevel section. Finer section grouping is obeyed, but no labels are created
or displayed.

The options frames scroll with the mouse wheel.

If possible the window opens wide enough to display its contents. This is
accomplished by adding a binding to the scrolled frames which report their
widths to their parent frames.

Programs using the label and description attributes on options or tabs will
now have the description displayed as balloon help.

The help button is not enabled if no help file exists for the program. (closes
wish 3799)

The type and required text for options is displayed right justified.

Module information is displayed differently. Both the label and description
are displayed if available. Module is displayed with an icon if a matching
one can be found in $(GISBASE)/etc/icons. See /gui/icons/README for more
information.

Buttons for file selection are displayed with an open file icon.

Buttons for selecting database elements and symbols are displayed with an icon
if a matching one can be found in $(GISBASE)/etc/icons.

It is now possible to determine when a command has finished executing. Output
text has an added graphic along with the command text when the command is
run. When the command finishes another graphic is displayed along with the
text "Done."