This is a fairly major bugfix combined with a minor code cleanup and feature upgrade. The bugfix addresses a serious issue of stability in the adaptive "soft-knee" smoothing algorithm. Prior to the fix, the adaptive smoothing process could get quickly out of control if only a few very large values were calulated, and thus the spike would eventually cease to be surpressed. The bugfix changes the adaptive smoothing process to look at descriptive statistics less affected by large values (namely percentiles and quartiles rather than means and standard deviations).
Code cleanup includes removal of several commented out sections of older code left in place during previous feature upgrade tests, removal of all code dealing with MASKS (there hasn't really been a need for that code in a long time), addition of extra detail in some code comments, and general spelling corrections.

Minor feature upgrades include change to the calcualtion of flow depth to make it more accurate and to make the flow depth cutoffs more realistic, and changes to some of the "default" values that appear in the g.parser gui. None of these changes will affect current usage of the script.