wxGUI: icons for layers in lmgr tree without 'layer' and 'add' symbols
Size of icons is 16x16 to avoid imperfect resampling at some (or all) platforms.
The motivation is to make the icons more useful, they now can actually help to recognize type of layer.
The commit assumes that the icon objects defined in layertree are used only for layers, not for other things although item label suggests that.
Command and grid layer still use the layer-xxx-add icon because there is no SVG for them in OSGeo icons (trac.osgeo.org/osgeo/browser/graphics/trunk/).
Geodesic and rhumbline layers are already using an icon with this style (the same icon, actually).