gunittest: do not import test file, support also sh files
Implemented quickly and simply on the cost of loosing support for unittest-like discovery and execution. It can be added back again using the old code.

Also the support for different location tests was dropped for now because parsing of the file to get the location type is missing (attribute access after import was used before). This has to be implemented anyway. However, for now support for scripts seems to be a more important feature.