gunittest: basic support for Python 2.6 and appropriate help for individual test file
Only import of modules is supposed to work with 2.6 to allow documentation build and perhaps running the machinery.

However, the individual tests may or may not work since a lot of assert methods and set up class step are missing and are impossible to provide without reimplementing unittest for 2.7. The full compatibility for 2.6 would be possible with unittest2 which even provides 3.2 features in 2.7 but this is a package which has to be installed separately.

The default help is fortunately the right one. The other help is not meant for individual files and is not possible to easily import for both 2.6 and 2.7 (and is not part of the API).