wxGUI/startup: change layout of startup/welcome window/screen
Using philosophy of editing lists (list of X with buttons to manage items). There are three things to select, so there are three boxes to organzie things. Each box has a description of what it is directly in the window, so there is no need to open manual in order to understand basic things. Delete and rename is done through buttons rather then choice/combo which, unlike button, should not triger any action.

Size of the window and inner elements might be good change. Window could be wider and minimal height lower.

The green title text in the window was removed. The version and revision are now in the window title.

Some changes to layout should be considered. Head picture might be too close to the first group and the message/warning/error area could go somewhere else (typicaly at top or bottom, top would be under head image in our case); now using the old one (under gisdbase).

For location and mapset using GRASS Location, Location, GRASS Mapset and Mapset. Alternative primary names as well as box labels (e.g. with numbers) and box descriptions should be easy to introduce by simple change of the strings (without a need to modify layout).

Start button is now using 'Start GRASS session' to expose the concept of session (one can start GUI multiple times, one GRASS session on gisdbase/loc/map combination, session must run to run Pyrthon scripts, ...).

Naming of GISDBASE is unclear, programming manual is using (GRASS) database but it seems more that location is the databse and mapset the namespace and it is a quite normal directory, so startup screen is using GRASS GIS data directory.