docs: manually created index with images
The generated index of all images is now 'manual gallery'. The graphical index is now manually created list with links to selected parts or, potentially, other manually created indexes.

Images are low res JPGs because the page must load quickly.

Images on page are relative to font size (em), this kind of relative size but not dependent on browser window size. More importantly, the items are in a list rather than a table, so they layout on the wide screens as wide and as one column on narrow displays.