don't change links for the actual HTML files, don't guess from addon names
r66479 and r66480 were doing nothing as the actual HTMLs are in docs/html directory.

Adding path to existing HTMLs as a 3rd parameter to update_manual.py. In this
way it is possible to test it locally.

The tree is traversed for every file to find existing HTMLs (as opposed to
having one script to fix all files or storing filenames in a file).
With all HTMLs in trunk's dist dir and searching in trunk including dist dir
it takes real 3m45.726s, user 2m16.512s, sys 1m23.124s. The addons case
should be smaller.