r.futures.pga: carry on the cells we have developed over the limit
The counting mechanism was in the original code but the value was not stored
in between the steps. This would probably work if the variable would be static
and we would do only one region.
    
The next step would be to warn at the end if there are some regions which
are developed too much, i.e. the value is not zero.