v.in.lidar: import points only in selected areas [news]
It uses a vector map with areas as a mask to specify
where to import the points (ignores points outside
the mask, does the opposite with -i flag).

Cashing the bboxes seems to be slightly faster
than just creating them all over again. Otherwise,
this implementation is the simplest (not fastest) way possible.